requirements.txt
aiobotocore
aiogram
aiohttp
aiozk
alembic
binance
funcy
joblib
kubernetes
matplotlib
numba
numpy==1.22.0
pandas
pandas_ta
plotly
psycopg2-binary
python-binance
python-telegram-bot
requests
s3cmd
sklearn
sortedcontainers
sqlalchemy
tensorflow